+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 [unreleased]
+* add filter to set low timeout cache to overcome FAIR caching and return current data
 
 #### 0.8.2 / 2026-02-12
 * composer update to GUL 2.8.3